+++ src/core/pp.c
@@ -629,7 +629,7 @@ void janet_buffer_format(
                 }
                 case 's': {
                     const uint8_t *s = janet_getstring(argv, arg);
-                    size_t l = janet_string_length(s);
+                    int32_t l = janet_string_length(s);
                     if (form[2] == '\0')
                         janet_buffer_push_bytes(b, s, l);
                     else {
